# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaaS ERP API",
        "description": "API Documentation for the Multi-tenant SaaS ERP Application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/v1/auth/tfa/enable": {
            "post": {
                "tags": [
                    "Two-Factor Auth"
                ],
                "summary": "Enable Two-Factor Authentication (Stage 1)",
                "description": "Enable Two-Factor Authentication (TFA) - Stage 1: Generate Google Authenticator Secret & QR code",
                "operationId": "e8fb83e6553445a4106667199ff0b370",
                "responses": {
                    "200": {
                        "description": "TFA Secret generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "secret": {
                                            "type": "string"
                                        },
                                        "qr_code_url": {
                                            "type": "string"
                                        },
                                        "current_otp": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/tfa/confirm": {
            "post": {
                "tags": [
                    "Two-Factor Auth"
                ],
                "summary": "Confirm Two-Factor Authentication (Stage 2)",
                "description": "Confirm TFA - Stage 2: Verify TOTP code and activate",
                "operationId": "d22d546b8d6891838d130e2a3a11ed58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "TFA Confirmed and Activated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "recovery_codes": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "TFA secret not generated"
                    },
                    "422": {
                        "description": "Invalid code"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/tfa/disable": {
            "post": {
                "tags": [
                    "Two-Factor Auth"
                ],
                "summary": "Disable Two-Factor Authentication",
                "description": "Disable TFA",
                "operationId": "3a87f3324a3bbce044926f2e3000408c",
                "responses": {
                    "200": {
                        "description": "TFA Disabled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/tfa/verify": {
            "post": {
                "tags": [
                    "Two-Factor Auth"
                ],
                "summary": "Verify TFA code during login challenge",
                "description": "Verify TOTP code during login challenge",
                "operationId": "bfd9ffa722f6259beced17c635d58a7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tfa_token",
                                    "code"
                                ],
                                "properties": {
                                    "tfa_token": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verification successful, returns tokens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User not configured for TFA"
                    },
                    "422": {
                        "description": "Invalid code or challenge session"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}